Chapter 6:4 - Replaced ImageLink with ImageLinks
Replaced it with ImageLinks with three images.. Removed modals from code, to get the links to work. Think more coding is needed in the course, to get the modals to work, so I've only commented out the code.
A modal (or modal window) is a small window that pops up on top of the page, usually dimming the background, and requires the user to interact with it before they can continue.